Fix icon sizes on mobile.
So on desktop it will choose 16x16, 32x32 or 48x48, on most screens.
Mobile probably has higher dpi and we don't have high res icons, so
scale.. otherwise icons will look too small on mobile.

Anyway, whatever the device is the icon will have the same size in cm/inches,
so it feels confortable for the finger.