OSX: Fix sticky mode.
Qt::Tool doesn't work very well, and we can't set Qt::WA_MacAlwaysShowToolWindow
because it's not a widget, but anyway, Qt::Tool is moot on OSX because it doesn't
affect the taskbar in this OS, so just use Qt::Window.

And on OSX Qt::WindowStaysOnTopHint doesn't work on Tool, only Window, so
this commit is perfect.