New feature: saveOrOpenUrl() DBus method
User will be asked if he wants to open the URL or add a new task with it.

You can now queue URLs you get in IM/e-mail instead of interrupting
your flow.

Just configure your IM client to use flow's DBus interface, this way
if you get a link with cat pictures, click it, choose "Save", and a new
task is created for you.